fixed compiler warning on windows

+++ Lua4RSTickThread.cpp
@@ -24,7 +24,7 @@ void Lua4RSTickThread::run()
     while(isRunning())
     {
         // tick each X second
-        if(_lastRun + tickIntervalInSeconds <= time(0) && _startUpEventTriggered)
+        if(_lastRun + tickIntervalInSeconds <= (uint)time(0) && _startUpEventTriggered)
         {
             LuaEvent e;
             e.eventId = L4R_TIMERTICK;
@@ -38,7 +38,7 @@ void Lua4RSTickThread::run()
         }
 
         // start up event
-        if(!_startUpEventTriggered && _initTime + secondsToStarUpEvent <= time(0))
+        if(!_startUpEventTriggered && (_initTime + secondsToStarUpEvent) <= (uint)time(0))
         {
             LuaEvent e;
             e.eventId = L4R_STARTUP;
